server_py/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.1
pydantic>=2.0
requests>=2.31.0
docling  # Replace with a version specifier if needed, e.g., docling>=0.1.0
pyjwt
openai==1.91.0
pinecone==7.2.0
tiktoken>=0.5.0